Move synthesizer_utils to contracts/utils
