+++ src/goto-instrument/contracts/utils.h
@@ -14,6 +14,7 @@ Date: September 2021
 #include <goto-programs/goto_convert_class.h>
 
 #include <goto-programs/goto_model.h>
+#include <goto-programs/loop_ids.h>
 
 #include <goto-instrument/havoc_utils.h>
 
@@ -23,6 +24,10 @@ Date: September 2021
 #define ENTERED_LOOP "__entered_loop"
 #define IN_LOOP_HAVOC_BLOCK "__in_loop_havoc_block"
 #define INIT_INVARIANT "__init_invariant"
+
+template <class T, typename C>
+class loop_templatet;
+typedef std::map<loop_idt, exprt> invariant_mapt;
 
 /// Class that allows to clean expressions of side effects and to generate
 /// havoc_slice expressions.
@@ -270,5 +275,46 @@ bool is_assignment_to_instrumented_variable(
 
 /// Convert the suffix digits right after `prefix` of `str` into unsigned.
 unsigned get_suffix_unsigned(const std::string &str, const std::string &prefix);
+
+/// Find the goto instruction of `loop` that jumps to `loop_head`
+goto_programt::targett get_loop_end_from_loop_head_and_content_mutable(
+  const goto_programt::targett &loop_head,
+  const loop_templatet<goto_programt::targett, goto_programt::target_less_than>
+    &loop);
+
+goto_programt::const_targett get_loop_end_from_loop_head_and_content(
+  const goto_programt::const_targett &loop_head,
+  const loop_templatet<
+    goto_programt::const_targett,
+    goto_programt::target_less_than> &loop);
+
+/// Return loop head if `finding_head` is true,
+/// Otherwise return loop end.
+goto_programt::targett get_loop_head_or_end(
+  const unsigned int loop_number,
+  goto_functiont &function,
+  bool finding_head);
+
+/// Find and return the last instruction of the natural loop with
+/// `loop_number` in `function`. loop_end -> loop_head
+goto_programt::targett
+get_loop_end(const unsigned int loop_number, goto_functiont &function);
+
+/// Find and return the first instruction of the natural loop with
+/// `loop_number` in `function`. loop_end -> loop_head
+goto_programt::targett
+get_loop_head(const unsigned int loop_number, goto_functiont &function);
+
+/// Annotate the invariants in `invariant_map` to their corresponding
+/// loops. Corresponding loops are specified by keys of `invariant_map`
+void annotate_invariants(
+  const invariant_mapt &invariant_map,
+  goto_modelt &goto_model);
+
+/// Annotate the assigns in `assigns_map` to their corresponding
+/// loops. Corresponding loops are specified by keys of `assigns_map`
+void annotate_assigns(
+  const std::map<loop_idt, std::set<exprt>> &assigns_map,
+  goto_modelt &goto_model);
 
 #endif // CPROVER_GOTO_INSTRUMENT_CONTRACTS_UTILS_H
